refactor(objectstorage): deduplicate enableProject into the utils package
Move the retrying enableProject helper from the bucket, credential and
credentialsgroup resources to objectstorage/utils as EnableProject, and
move its tests to utils/util_test.go so the retry behaviour is covered
once for every caller.
