+++ stackit/internal/services/objectstorage/bucket/resource.go
@@ -6,7 +6,6 @@ import (
 	"fmt"
 	"net/http"
 	"strings"
-	"time"
 
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/booldefault"
 	"github.com/hashicorp/terraform-plugin-framework/resource/schema/boolplanmodifier"
@@ -212,7 +211,7 @@ func (r *bucketResource) Create(ctx context.Context, req resource.CreateRequest,
 	ctx = tflog.SetField(ctx, "region", region)
 
 	// Handle project init
-	err := enableProject(ctx, &model, region, r.client.DefaultAPI)
+	err := objectstorageUtils.EnableProject(ctx, projectId, region, r.client.DefaultAPI)
 	if err != nil {
 		core.LogAndAddError(ctx, &resp.Diagnostics, "Error creating bucket", fmt.Sprintf("Enabling object storage project before creation: %v", err))
 		return
@@ -392,33 +391,5 @@ func mapFields(bucketResp *objectstorage.GetBucketResponse, model *Model, region
 	model.URLVirtualHostedStyle = types.StringValue(bucket.UrlVirtualHostedStyle)
 	model.Region = types.StringValue(region)
 	model.ObjectLock = types.BoolValue(bucket.ObjectLockEnabled)
-	return nil
-}
-
-const (
-	// Two object storage resources created in the same apply enable the project concurrently;
-	// the API answers the losing call with 409. See enableProject.
-	enableProjectAttempts = 4
-)
-
-// Overridden in tests to keep them fast.
-var enableProjectRetryDelay = 2 * time.Second
-
-// enableProject enables object storage for the specified project. If the project is already enabled, nothing happens
-func enableProject(ctx context.Context, model *Model, region string, client objectstorage.DefaultAPI) error {
-	projectId := model.ProjectId.ValueString()
-
-	// From the object storage OAS: Creation will also be successful if the project is already enabled, but will not create a duplicate.
-	// That holds for sequential calls. Two object storage resources created in the same apply call this concurrently,
-	// and the API rejects the second one with 409 project.create_conflict ("Two concurrent calls try to create the
-	// same project"). Retrying is safe: once the competing call has finished, enabling an already enabled project succeeds.
-	config := utils.RetryConfig{
-		Attempts:         enableProjectAttempts,
-		Delay:            enableProjectRetryDelay,
-		RetryStatusCodes: []int{http.StatusConflict},
-	}
-	if _, err := utils.RetryRequest(ctx, client.EnableService(ctx, projectId, region).Execute, config); err != nil {
-		return fmt.Errorf("failed to create object storage project: %w", err)
-	}
 	return nil
 }
