+++ stackit/internal/services/objectstorage/credential/resource_test.go
@@ -15,22 +15,6 @@ import (
 	objectstorage "github.com/stackitcloud/stackit-sdk-go/services/objectstorage/v2api"
 )
 
-type mockSettings struct {
-	returnError bool
-}
-
-func newAPIMock(settings *mockSettings) objectstorage.DefaultAPI {
-	return &objectstorage.DefaultAPIServiceMock{
-		EnableServiceExecuteMock: new(func(_ objectstorage.ApiEnableServiceRequest) (*objectstorage.ProjectStatus, error) {
-			if settings.returnError {
-				return nil, fmt.Errorf("create project failed")
-			}
-
-			return &objectstorage.ProjectStatus{}, nil
-		}),
-	}
-}
-
 func TestMapFields(t *testing.T) {
 	now := time.Now()
 	const testRegion = "eu01"
@@ -155,79 +139,6 @@ func TestMapFields(t *testing.T) {
 				if diff != "" {
 					t.Fatalf("Data does not match: %s", diff)
 				}
-			}
-		})
-	}
-}
-
-func TestEnableProject(t *testing.T) {
-	// enableProject retries, and the mock returns a plain error rather than an
-	// *oapierror.GenericOpenAPIError - RetryRequest only filters by status code
-	// when it can type-assert the error, so the failing case uses up every
-	// attempt. Without shrinking the delay this test would sleep for seconds.
-	oldDelay := enableProjectRetryDelay
-	enableProjectRetryDelay = time.Millisecond
-	defer func() { enableProjectRetryDelay = oldDelay }()
-
-	const testRegion = "eu01"
-	id := fmt.Sprintf("%s,%s,%s", "pid", testRegion, "cgid,cid")
-	tests := []struct {
-		description string
-		expected    Model
-		enableFails bool
-		isValid     bool
-	}{
-		{
-			"default_values",
-			Model{
-				Id:                  types.StringValue(id),
-				ProjectId:           types.StringValue("pid"),
-				CredentialsGroupId:  types.StringValue("cgid"),
-				CredentialId:        types.StringValue("cid"),
-				Name:                types.StringNull(),
-				AccessKey:           types.StringNull(),
-				SecretAccessKey:     types.StringNull(),
-				ExpirationTimestamp: types.StringNull(),
-				RotateWhenChanged:   types.MapNull(types.StringType),
-			},
-			false,
-			true,
-		},
-		{
-			"error_response",
-			Model{
-				Id:                  types.StringValue(id),
-				ProjectId:           types.StringValue("pid"),
-				CredentialsGroupId:  types.StringValue("cgid"),
-				CredentialId:        types.StringValue("cid"),
-				Name:                types.StringNull(),
-				AccessKey:           types.StringNull(),
-				SecretAccessKey:     types.StringNull(),
-				ExpirationTimestamp: types.StringNull(),
-				RotateWhenChanged:   types.MapNull(types.StringType),
-			},
-			true,
-			false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.description, func(t *testing.T) {
-			client := newAPIMock(&mockSettings{
-				returnError: tt.enableFails,
-			})
-
-			model := &Model{
-				ProjectId:          tt.expected.ProjectId,
-				CredentialsGroupId: tt.expected.CredentialsGroupId,
-				CredentialId:       tt.expected.CredentialId,
-				RotateWhenChanged:  types.MapNull(types.StringType),
-			}
-			err := enableProject(context.Background(), model, "eu01", client)
-			if !tt.isValid && err == nil {
-				t.Fatalf("Should have failed")
-			}
-			if tt.isValid && err != nil {
-				t.Fatalf("Should not have failed: %v", err)
 			}
 		})
 	}
