+++ stackit/internal/services/objectstorage/credentialsgroup/resource_test.go
@@ -3,11 +3,7 @@ package objectstorage
 import (
 	"context"
 	"fmt"
-	"net/http"
 	"testing"
-	"time"
-
-	"github.com/stackitcloud/stackit-sdk-go/core/oapierror"
 
 	"github.com/google/go-cmp/cmp"
 	"github.com/hashicorp/terraform-plugin-framework/types"
@@ -21,13 +17,6 @@ type mockSettings struct {
 
 func newAPIMock(settings *mockSettings) objectstorage.DefaultAPI {
 	return &objectstorage.DefaultAPIServiceMock{
-		EnableServiceExecuteMock: new(func(_ objectstorage.ApiEnableServiceRequest) (*objectstorage.ProjectStatus, error) {
-			if settings.returnError {
-				return nil, fmt.Errorf("create project failed")
-			}
-
-			return &objectstorage.ProjectStatus{}, nil
-		}),
 		ListCredentialsGroupsExecuteMock: new(func(_ objectstorage.ApiListCredentialsGroupsRequest) (*objectstorage.ListCredentialsGroupsResponse, error) {
 			if settings.returnError {
 				return nil, fmt.Errorf("get credentials groups failed")
@@ -129,48 +118,6 @@ func TestMapFields(t *testing.T) {
 				if diff != "" {
 					t.Fatalf("Data does not match: %s", diff)
 				}
-			}
-		})
-	}
-}
-
-func TestEnableProject(t *testing.T) {
-	// enableProject retries, and the mock returns a plain error rather than an
-	// *oapierror.GenericOpenAPIError - RetryRequest only filters by status code
-	// when it can type-assert the error, so the failing case uses up every
-	// attempt. Without shrinking the delay this test would sleep for seconds.
-	oldDelay := enableProjectRetryDelay
-	enableProjectRetryDelay = time.Millisecond
-	defer func() { enableProjectRetryDelay = oldDelay }()
-
-	tests := []struct {
-		description string
-		enableFails bool
-		isValid     bool
-	}{
-		{
-			"default_values",
-			false,
-			true,
-		},
-		{
-			"error_response",
-			true,
-			false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.description, func(t *testing.T) {
-			client := newAPIMock(&mockSettings{
-				returnError: tt.enableFails,
-			})
-
-			err := enableProject(context.Background(), &Model{}, "eu01", client)
-			if !tt.isValid && err == nil {
-				t.Fatalf("Should have failed")
-			}
-			if tt.isValid && err != nil {
-				t.Fatalf("Should not have failed: %v", err)
 			}
 		})
 	}
@@ -327,72 +274,5 @@ func TestReadCredentialsGroups(t *testing.T) {
 				}
 			}
 		})
-	}
-}
-
-// Two object storage resources created in the same apply enable the project concurrently.
-// The API answers the losing call with 409 project.create_conflict; enableProject must retry
-// instead of failing the apply.
-func TestEnableProjectRetriesOnConflict(t *testing.T) {
-	tests := []struct {
-		description  string
-		conflicts    int
-		isValid      bool
-		wantAttempts int
-	}{
-		{"succeeds immediately", 0, true, 1},
-		{"one conflict, then success", 1, true, 2},
-		{"conflicts until the attempts are used up", enableProjectAttempts, false, enableProjectAttempts},
-	}
-
-	old := enableProjectRetryDelay
-	enableProjectRetryDelay = time.Millisecond
-	defer func() { enableProjectRetryDelay = old }()
-
-	for _, tt := range tests {
-		t.Run(tt.description, func(t *testing.T) {
-			attempts := 0
-			client := &objectstorage.DefaultAPIServiceMock{
-				EnableServiceExecuteMock: new(func(_ objectstorage.ApiEnableServiceRequest) (*objectstorage.ProjectStatus, error) {
-					attempts++
-					if attempts <= tt.conflicts {
-						return nil, &oapierror.GenericOpenAPIError{StatusCode: http.StatusConflict}
-					}
-					return &objectstorage.ProjectStatus{}, nil
-				}),
-			}
-
-			ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-			defer cancel()
-
-			err := enableProject(ctx, &Model{}, "eu01", client)
-			if tt.isValid && err != nil {
-				t.Fatalf("Should not have failed: %v", err)
-			}
-			if !tt.isValid && err == nil {
-				t.Fatal("Should have failed")
-			}
-			if attempts != tt.wantAttempts {
-				t.Fatalf("Expected %d attempts, got %d", tt.wantAttempts, attempts)
-			}
-		})
-	}
-}
-
-// A non-conflict error must not be retried.
-func TestEnableProjectDoesNotRetryOtherErrors(t *testing.T) {
-	attempts := 0
-	client := &objectstorage.DefaultAPIServiceMock{
-		EnableServiceExecuteMock: new(func(_ objectstorage.ApiEnableServiceRequest) (*objectstorage.ProjectStatus, error) {
-			attempts++
-			return nil, &oapierror.GenericOpenAPIError{StatusCode: http.StatusForbidden}
-		}),
-	}
-
-	if err := enableProject(context.Background(), &Model{}, "eu01", client); err == nil {
-		t.Fatal("Should have failed")
-	}
-	if attempts != 1 {
-		t.Fatalf("Expected a single attempt, got %d", attempts)
 	}
 }
